More cleanup + documenting
- documenting the routes and removed a lot of the ones that are no longer used
