+++ main/server/routes/years.js
@@ -1,27 +1,7 @@
 const yearsRouter = require("express").Router();
 const { Year } = require("../models/year");
-const { Month } = require("../models/month");
-const { Envelope } = require("../models/envelope");
+const { createYear } = require("../controllers/year");
 
-async function createYear(info) {
-  if (info.months === undefined) throw new Error("Not enough information provided");
-  const months = Object.keys(info.months).map((monthInfo) => {
-    const newMonth = new Month({
-      month: monthInfo,
-      budget: info.months[monthInfo].budget,
-      remaining: info.months[monthInfo].budget,
-    });
-    return newMonth;
-  });
-  const year = new Year({
-    year: info.year,
-    budget: info.budget,
-    remaining: info.budget,
-    months,
-  });
-  await year.save();
-  return year;
-}
 /** POST ROUTES */
 /**
  * Create a new year
@@ -36,49 +16,9 @@ yearsRouter.post("/createYear", async (req, res) => {
   }
 });
 
-/**
- * Create a new envelope for a specific month
- * @param {number} year - year to get
- * @param {string} month - name of month
- * @param {object} body - monthId (needs to be _id of month), category (any string), budget (number)
- */
-yearsRouter.post("/:year/:month/add", async (req, res) => {
-  try {
-    const { monthId, category, budget } = req.body;
-    const envelope = new Envelope({
-      category: category,
-      budget: budget,
-      monthId,
-    });
-    await envelope.save();
-    const incAmount = Number(budget);
-    const decAmount = incAmount * -1;
-    const updateDocument = {
-      $push: { "months.$[months].envelopes": envelope._id },
-      $inc: {
-        remaining: decAmount,
-        spent: incAmount,
-        "months.$[months].remaining": decAmount,
-        "months.$[months].spent": incAmount,
-      },
-    };
-    const options = {
-      arrayFilters: [
-        {
-          "months._id": monthId,
-        },
-      ],
-    };
-    await Year.findOneAndUpdate({ year: req.params.year }, updateDocument, options);
-    res.status(200).json();
-  } catch (e) {
-    res.status(400).json({ message: e.message });
-  }
-});
-
 /** GET ROUTES */
 /**
- * Get all year documents
+ * @description Get all year documents
  */
 yearsRouter.get("/", async (req, res) => {
   try {
@@ -90,7 +30,7 @@ yearsRouter.get("/", async (req, res) => {
 });
 
 /**
- * Get year by number
+ * @description Get year by number
  */
 yearsRouter.get("/:year", async (req, res) => {
   try {
@@ -103,7 +43,7 @@ yearsRouter.get("/:year", async (req, res) => {
 });
 
 /**
- * Get months of a year
+ * @description Get all months of a year
  */
 yearsRouter.get("/:year/months", async (req, res) => {
   try {
@@ -116,7 +56,7 @@ yearsRouter.get("/:year/months", async (req, res) => {
 });
 
 /**
- * Get month of a year
+ * @description Get month of a year
  */
 yearsRouter.get("/:year/:month", async (req, res) => {
   try {
@@ -131,38 +71,9 @@ yearsRouter.get("/:year/:month", async (req, res) => {
   }
 });
 
-/**
- * Get all envelopes id's of a month of a year
- */
-yearsRouter.get("/:year/:month/allId", async (req, res) => {
-  try {
-    const { year, month } = req.params;
-    const projection = {
-      months: { $elemMatch: { month: month } },
-    };
-    const data = await Year.findOne({ year: year }, projection);
-    res.json(data.months[0].envelopes);
-  } catch (e) {
-    res.status(500).json({ message: e.message });
-  }
-});
-
-/**
- * Get all envelopes of a month of a year
- */
-yearsRouter.get("/:year/:monthId/all", async (req, res) => {
-  try {
-    const { monthId } = req.params;
-    const data = await Envelope.find({ month: monthId });
-    res.json(data);
-  } catch (e) {
-    res.status(500).json({ message: e.message });
-  }
-});
-
 /** PATCH ROUTES */
 /**
- * Update select year
+ * @description Update select year
  */
 yearsRouter.patch("/updateYear/:year", async (req, res) => {
   try {
@@ -187,7 +98,8 @@ yearsRouter.patch("/updateYear/:year", async (req, res) => {
 });
 
 /**
- * Update all months of select year
+ * @deprecated - think this can be removed but leaving it for now
+ * @description Update all months of select year
  */
 yearsRouter.patch("/:year/all", async (req, res) => {
   try {
@@ -204,6 +116,10 @@ yearsRouter.patch("/:year/all", async (req, res) => {
   }
 });
 
+/**
+ * @deprecated - think this can be removed but leaving it for now
+ * @description Update month of select year
+ */
 yearsRouter.patch("/:year/:month", async (req, res) => {
   try {
     const updateDocument = {
@@ -229,7 +145,7 @@ yearsRouter.patch("/:year/:month", async (req, res) => {
 
 /** DELETE ROUTES */
 /**
- * Delete all year documents
+ * @description Delete all year documents
  */
 yearsRouter.delete("/all", async (req, res) => {
   try {
@@ -241,44 +157,12 @@ yearsRouter.delete("/all", async (req, res) => {
 });
 
 /**
- * Delete year by number
+ * @description Delete year
  */
 yearsRouter.delete("/deleteYear/:year", async (req, res) => {
   try {
     const year = req.params.year;
     const data = await Year.deleteOne({ year: year });
-    res.json(data);
-  } catch (e) {
-    res.status(500).json({ message: e.message });
-  }
-});
-
-/**
- * Delete envelope from month by id
- */
-yearsRouter.delete("/:year/:month/:id", async (req, res) => {
-  try {
-    const { year, month, id } = req.params;
-    const envelope = await Envelope.findOneAndDelete({ _id: id });
-    const incAmount = envelope.budget;
-    const decAmount = envelope.budget * -1;
-    const updateDocument = {
-      $pull: { "months.$[months].envelopes": id },
-      $inc: {
-        remaining: incAmount,
-        spent: decAmount,
-        "months.$[months].remaining": incAmount,
-        "months.$[months].spent": decAmount,
-      },
-    };
-    const options = {
-      arrayFilters: [
-        {
-          "months.month": month,
-        },
-      ],
-    };
-    const data = await Year.findOneAndUpdate({ year: year }, updateDocument, options);
     res.json(data);
   } catch (e) {
     res.status(500).json({ message: e.message });
